+++ lib/packages/wikibase/data-model/src/Statement/StatementList.php
@@ -12,7 +12,6 @@ use Wikibase\DataModel\Claim\Claim;
 use Wikibase\DataModel\Entity\PropertyId;
 use Wikibase\DataModel\Reference;
 use Wikibase\DataModel\ReferenceList;
-use Wikibase\DataModel\References;
 use Wikibase\DataModel\Snak\Snak;
 use Wikibase\DataModel\Snak\SnakList;
 use Wikibase\DataModel\Snak\Snaks;
@@ -97,7 +96,7 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 	/**
 	 * @param Snak $mainSnak
 	 * @param Snak[]|Snaks|null $qualifiers
-	 * @param Reference[]|References|null $references
+	 * @param Reference[]|ReferenceList|null $references
 	 * @param string|null $guid
 	 */
 	public function addNewStatement( Snak $mainSnak, $qualifiers = null, $references = null, $guid = null ) {