+++ lib/packages/wikibase/data-model/RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * Removed all class aliases directly in the `Wikibase` namespace (deprecated since 0.6 or before)
 * Removed the class aliases for `ItemDiff` and `EntityDiff` (deprecated since 1.0)
 * Removed support for deserializing `EntityId` instances from version 0.4 or earlier
+* Removed `References` interface in favour of `ReferenceList`
 
 ## Version 1.1 (2014-09-29)